Adds automate, enable and gather requests.  moves the worker, tool, foreman add to use the original totalresource one.
